Rename trip folder to tripview
