+++ NativeAppsII_Windows_Groep18/View/TripView/AddTripView.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="NativeAppsII_Windows_Groep18.View.Trip.AddTripView"
+    x:Class="NativeAppsII_Windows_Groep18.View.TripView.AddTripView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:NativeAppsII_Windows_Groep18.View"
