+++ NativeAppsII_Windows_Groep18/View/TripView/TripOverview.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="NativeAppsII_Windows_Groep18.View.Trip.TripOverview"
+    x:Class="NativeAppsII_Windows_Groep18.View.TripView.TripOverview"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
@@ -26,9 +26,10 @@
                   IsItemClickEnabled="True">
                 <GridView.ItemTemplate>
                     <DataTemplate x:DataType="model:Trip">
-                        <Grid>
                             <Grid>
-                                <StackPanel>
+                                <StackPanel x:Name="TripGridViewItem"
+                                            Margin="24"
+                                            Tapped="NavigateToDetail">
                                     <Grid>
                                         <Grid.RowDefinitions>
                                             <RowDefinition Height="Auto" />
@@ -52,10 +53,9 @@
                                                Height="1"
                                                HorizontalAlignment="Stretch"
                                                VerticalAlignment="Bottom"
-                                               />
+                                               Fill="Black"/>
                                 </StackPanel>
                             </Grid>
-                        </Grid>
                     </DataTemplate>
                 </GridView.ItemTemplate>
             </GridView>
