+++ NativeAppsII_Windows_Groep18/View/TripView/TripOverview.xaml.cs
@@ -2,10 +2,12 @@
 using Microsoft.Extensions.DependencyInjection;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
+using Windows.UI.Xaml.Input;
+using NativeAppsII_Windows_Groep18.Model;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace NativeAppsII_Windows_Groep18.View.Trip
+namespace NativeAppsII_Windows_Groep18.View.TripView
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
@@ -32,6 +34,12 @@ namespace NativeAppsII_Windows_Groep18.View.Trip
             {
                 TripViewModel.LoadTrips();
             }
+        }
+
+        private void NavigateToDetail(object sender, TappedRoutedEventArgs e)
+        {
+            if (TripGridView.SelectedItem is Trip clickedItem)
+                Frame.Navigate(typeof(TripDetail), clickedItem);
         }
         #endregion
     }
